Add AutoSD platform adapters, H755ZI-Q board fixes, and CI pipeline - Implement real Kuksa gRPC and NATS cloud transport adapters with   systemd lifecycle integration for AutoSD HPC platform - Fix Nucleo-H755ZI-Q board overlay: add led1/led2 aliases, static IP   config, and disable ADC (Zephyr DTS binding update) - Add RPM packaging with Packit CI, Containerfile, and systemd service - Add GitHub Actions CI workflow for host and unit test builds - Add Renode ADC peripheral model and integration test scaffolding - Add protobuf/gRPC schema for Kuksa signal interface

+++ libs/body/speed-simulator/src/SpeedSimulator.cpp
@@ -34,8 +34,9 @@ void SpeedSimulator::init() {
     timer_id_ = timer_.startPeriodic(config_.update_interval_ms,
                                      [this]() { onTimerTick(); });
 
-    printk("[speed_sim] Initialized (interval=%ums, max=%.0f km/h)\n",
-           config_.update_interval_ms, config_.max_speed_kmh);
+    printk("[speed_sim] Initialized (interval=%ums, max=%d km/h)\n",
+           config_.update_interval_ms,
+           static_cast<int>(config_.max_speed_kmh));
 }
 
 void SpeedSimulator::onTimerTick() {
